refactor: Remove unused models & move models
This removes some legacy models that are no longer used
due to the new structure. In addition, all models related to
revisions are moved to the settings models.
